requirements.txt
###### Requirements without Version Specifiers ######`
python-dotenv
json
requests
psycopg2

###### Requirements with Version Specifiers ######`
